+++ src/uni_modules/wot-design-uni/components/wd-badge/wd-badge.vue
@@ -1,14 +1,14 @@
 <!--
  * @Author: weisheng
  * @Date: 2023-06-12 18:40:59
- * @LastEditTime: 2023-08-22 23:23:37
+ * @LastEditTime: 2023-10-10 16:38:30
  * @LastEditors: weisheng
  * @Description: 
  * @FilePath: \wot-design-uni\src\uni_modules\wot-design-uni\components\wd-badge\wd-badge.vue
  * 记得注释
 -->
 <template>
-  <view :class="['wd-badge', customClass]">
+  <view :class="['wd-badge', customClass]" :style="customStyle">
     <slot></slot>
     <view
       v-if="!hidden && (content || content === 0 || isDot)"
@@ -42,9 +42,11 @@ interface Props {
   top?: number
   right?: number
   customClass?: string
+  customStyle?: string
 }
 const props = withDefaults(defineProps<Props>(), {
   customClass: '',
+  customStyle: '',
   modelValue: null
 })
 const content = ref<number | string | null>(null)
@@ -66,7 +68,7 @@ function notice() {
   let value = props.modelValue
   const max = props.max
   if (value && max && typeof value === 'number' && !Number.isNaN(value) && !Number.isNaN(max)) {
-    value = max < value ? max : value
+    value = max < value ? `${max}+` : value
   }
   content.value = value
 }