+++ src/uni_modules/wot-design-uni/components/wd-datetime-picker/wd-datetime-picker.vue
@@ -137,7 +137,7 @@ export default {
 <script lang="ts" setup>
 import { getCurrentInstance, onBeforeMount, onMounted, ref, watch } from 'vue'
 import { deepClone, getType, isArray, isDef, isEqual, padZero } from '../common/util'
-import { useCell } from '../mixins/useCell'
+import { useCell } from '../composables/useCell'
 import { type DateTimeType, getPickerValue } from '../wd-datetime-picker-view/type'
 interface Props {
   customClass?: string
@@ -296,7 +296,7 @@ watch(
       // 每次value更新时都需要刷新整个列表
       innerValue.value = deepClone(getDefaultInnerValue())
     }
-    setShowValue()
+    setShowValue(false, false, true)
   },
   {
     deep: true,