+++ src/uni_modules/wot-design-uni/components/wd-popover/wd-popover.vue
@@ -14,36 +14,34 @@
         </view>
       </view>
     </view>
-    <wd-transition :show="modelValue" name="fade" :duration="200">
-      <view class="wd-popover__pos" :style="popover.popStyle.value">
-        <view :class="`wd-popover__container ${customPop}`">
-          <view
-            v-if="props.visibleArrow"
-            :class="`wd-popover__arrow ${popover.arrowClass.value} ${customArrow}`"
-            :style="popover.arrowStyle.value"
-          ></view>
-          <!-- 普通模式 -->
-          <view v-if="!useContentSlot && mode === 'normal'" class="wd-popover__inner">
-            {{ content }}
-          </view>
-          <!-- 列表模式 -->
-          <view v-if="!useContentSlot && mode === 'menu'" class="wd-popover__menu">
-            <view
-              v-for="(item, index) in content"
-              :key="index"
-              class="wd-popover__menu-inner"
-              @click="menuClick(index)"
-              :style="index === 0 ? 'border-top: none' : ''"
-            >
-              <wd-icon v-if="typeof item === 'object' && item.iconClass" :name="item.iconClass" custom-class="wd-popover__icon" />
-              <view style="display: inline-block">{{ typeof item === 'object' && item.content ? item.content : '' }}</view>
-            </view>
-          </view>
-          <!-- 用户自定义样式 -->
-          <slot name="content" v-else />
+    <wd-transition custom-class="wd-popover__pos" :custom-style="popover.popStyle.value" :show="modelValue" name="fade" :duration="200">
+      <view :class="`wd-popover__container ${customPop}`">
+        <view
+          v-if="props.visibleArrow"
+          :class="`wd-popover__arrow ${popover.arrowClass.value} ${customArrow}`"
+          :style="popover.arrowStyle.value"
+        ></view>
+        <!-- 普通模式 -->
+        <view v-if="!useContentSlot && mode === 'normal'" class="wd-popover__inner">
+          {{ content }}
         </view>
-        <wd-icon v-if="showClose" name="close" custom-class="wd-popover__close-icon" @click="toggle"></wd-icon>
+        <!-- 列表模式 -->
+        <view v-if="!useContentSlot && mode === 'menu'" class="wd-popover__menu">
+          <view
+            v-for="(item, index) in content"
+            :key="index"
+            class="wd-popover__menu-inner"
+            @click="menuClick(index)"
+            :style="index === 0 ? 'border-top: none' : ''"
+          >
+            <wd-icon v-if="typeof item === 'object' && item.iconClass" :name="item.iconClass" custom-class="wd-popover__icon" />
+            <view style="display: inline-block">{{ typeof item === 'object' && item.content ? item.content : '' }}</view>
+          </view>
+        </view>
+        <!-- 用户自定义样式 -->
+        <slot name="content" v-else />
       </view>
+      <wd-icon v-if="showClose" name="close" custom-class="wd-popover__close-icon" @click="toggle"></wd-icon>
     </wd-transition>
     <view @click="toggle" class="wd-popover__target" id="target">
       <slot />
@@ -63,9 +61,10 @@ export default {
 </script>
 
 <script lang="ts" setup>
-import { getCurrentInstance, onBeforeMount, onBeforeUnmount, onMounted, watch } from 'vue'
-import { usePopover } from '../mixins/usePopover'
+import { getCurrentInstance, inject, onBeforeMount, onBeforeUnmount, onMounted, watch } from 'vue'
+import { usePopover } from '../composables/usePopover'
 import { closeOther, pushToQueue, removeFromQueue } from '../common/clickoutside'
+import { type Queue, queueKey } from '../composables/useQueue'
 
 type PlacementType =
   | 'top'
@@ -114,6 +113,8 @@ const props = withDefaults(defineProps<Props>(), {
   mode: 'normal'
 })
 
+const queue = inject<Queue | null>(queueKey, null)
+
 const selector: string = 'popover'
 const emit = defineEmits(['update:modelValue', 'menuclick', 'change', 'open', 'close'])
 const { proxy } = getCurrentInstance() as any
@@ -140,7 +141,11 @@ watch(
   (newValue) => {
     if (newValue) {
       popover.control(props.placement, props.offset)
-      closeOther(proxy)
+      if (queue && queue.closeOther) {
+        queue.closeOther(proxy)
+      } else {
+        closeOther(proxy)
+      }
     }
     popover.showStyle.value = newValue ? 'display: inline-block;' : 'display: none;'
     emit('change', { show: newValue })
@@ -153,12 +158,20 @@ onMounted(() => {
 })
 
 onBeforeMount(() => {
-  pushToQueue(proxy)
+  if (queue && queue.pushToQueue) {
+    queue.pushToQueue(proxy)
+  } else {
+    pushToQueue(proxy)
+  }
   popover.showStyle.value = props.modelValue ? 'opacity: 1;' : 'opacity: 0;'
 })
 
 onBeforeUnmount(() => {
-  removeFromQueue(proxy)
+  if (queue && queue.removeFromQueue) {
+    queue.removeFromQueue(proxy)
+  } else {
+    removeFromQueue(proxy)
+  }
 })
 
 const popover = usePopover()